% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_new_submissions.R
\name{mod_new_submissions_ui}
\alias{mod_new_submissions_ui}
\title{mod_new_submissions_ui and mod_new_submissions_server}
\usage{
mod_new_submissions_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
