% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary_snapshot.R
\name{mod_summary_snapshot_ui}
\alias{mod_summary_snapshot_ui}
\alias{mod_summary_snapshot_server}
\title{mod_summary_snapshot_ui and mod_summary_snapshot_server}
\usage{
mod_summary_snapshot_ui(id)

mod_summary_snapshot_server(input, output, session, group_object, data_config)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
